DRY postUserMessage function
