+++ front/lib/api/assistant/streaming/helpers.ts
@@ -1,3 +1,11 @@
 export function getAgentExecutionChannelId(agentMessageId: string) {
   return `agent-execution-${agentMessageId}`;
 }
+
+export function getConversationChannelId({
+  conversationId,
+}: {
+  conversationId: string;
+}) {
+  return `conversation-${conversationId}`;
+}
